+++ modules/apps/analytics/analytics-reports-js-components-web/src/main/resources/META-INF/resources/js/index.ts
@@ -5,6 +5,7 @@
 
 import CMSPerformance from './CMSPerformance';
 import ContentDashboardPerformance from './ContentDashboardPerformance';
+import ConnectToAnalyticsCloud from './components/cms/ConnectToAnalyticsCloud';
 
 export * from './utils/buildQueryString';
 export * from './utils/date';
@@ -12,4 +13,4 @@ export * from './utils/math';
 export * from './components/RangeSelectorsDropdown';
 export * from './utils/metrics';
 
-export {ContentDashboardPerformance, CMSPerformance};
+export {CMSPerformance, ConnectToAnalyticsCloud, ContentDashboardPerformance};
